+++ lib/src/options.dart
@@ -47,6 +47,13 @@ class Options {
   )
   final bool version;
 
+  @CliOption(
+    abbr: 'w',
+    help: 'Include all packages in the workspace.',
+    negatable: false,
+  )
+  final bool workspace;
+
   @CliOption(abbr: '?', help: 'Print this help content.', negatable: false)
   final bool help;
 
@@ -61,6 +68,7 @@ class Options {
     required this.help,
     this.command,
     required this.version,
+    required this.workspace,
   }) : ignorePackages = ignorePackages ?? const [];
 }