+++ lib/src/viz_root.dart
@@ -23,6 +23,7 @@ class VizRoot {
     Iterable<String>? ignorePackages,
     bool directDependenciesOnly = false,
     bool productionDependenciesOnly = false,
+    bool includeWorkspace = false,
   }) async {
     final rootPubspec = service.rootPubspec();
 
@@ -30,9 +31,11 @@ class VizRoot {
       flagOutdated,
       directDependenciesOnly,
       productionDependenciesOnly,
+      includeWorkspace: includeWorkspace,
     );
 
-    final value = VizRoot._(rootPubspec.name, packages).._update();
+    final value = VizRoot._(rootPubspec.name, packages)
+      .._update(includeWorkspace);
 
     if (flagOutdated) {
       for (var dep in _allDeps(value, ignorePackages)) {
@@ -66,19 +69,25 @@ class VizRoot {
     return value;
   }
 
-  void _update() {
-    if (root.isPrimary == false) {
-      root.isPrimary = true;
+  void _update(bool includeWorkspace) {
+    // Collect primary packages at the start.
+    // If none are set (normal case), mark root as primary.
+    if (packages.values.every((v) => !v.isPrimary)) {
+      packages[rootPackageName]!.isPrimary = true;
+    }
 
-      assert(root.onlyDev);
-      root.onlyDev = false;
+    final primaryPackages = packages.values.where((v) => v.isPrimary).toList();
 
-      for (var primaryDep in root.dependencies) {
+    for (var pkg in primaryPackages) {
+      pkg.onlyDev = false;
+
+      for (var primaryDep in pkg.dependencies) {
         final package = packages[primaryDep.name];
         if (package == null) continue;
 
-        assert(!package.isPrimary);
-        package.isPrimary = true;
+        if (!includeWorkspace) {
+          package.isPrimary = true;
+        }
 
         if (!primaryDep.isDevDependency) {
           _updateDevOnly(primaryDep);